feat: stabilize vegetation instance generation
Replace parallel candidate lists with deterministic per-cell arrays and masks, allowing commit processing to use stable cell indices instead of candidate ordering. Expand vegetation IDs to `ulong` and derive them from global cells for stable cross-chunk identity, while simplifying disposal and preserving occupancy-based species prioritization.

+++ Assets/_Project/Features/ProceduralWorld/Domain/Vegetation/VegetationInstanceData.cs
@@ -23,8 +23,9 @@ namespace _Project.Features.ProceduralWorld.Domain.Vegetation
         public float Scale;
 
         /// <summary>
-        /// Stable identifier assigned to the generated instance.
+        /// Stable deterministic identifier derived from the global cell.
+        /// Species is part of the full identity because instances are grouped by species layer.
         /// </summary>
-        public uint Id;
+        public ulong Id;
     }
 }
